[Core] Bring back engine test that makes sure memory can be allocated outside a scope.
