+++ tfjs-core/src/engine_test.ts
@@ -23,6 +23,7 @@ import {ALL_ENVS, describeWithFlags, TestKernelBackend} from './jasmine_util';
 import {TensorInfo} from './kernel_registry';
 import {Tensor} from './tensor';
 import {expectArraysClose} from './test_util';
+import {BackendValues, DataType} from './types';
 
 describe('Backend registration', () => {
   beforeAll(() => {
@@ -695,5 +696,35 @@ describeWithFlags('Detects memory leaks in kernels', ALL_ENVS, () => {
     tf.removeBackend(backendName);
     tf.unregisterKernel(kernelName, backendName);
     tf.unregisterKernel(kernelNameComplex, backendName);
+  });
+});
+
+// NOTE: This describe is purposefully not a describeWithFlags so that we
+// test tensor allocation where no scopes have been created.
+describe('Memory allocation outside a test scope', () => {
+  it('constructing a tensor works', async () => {
+    const backendName = 'test-backend';
+    tf.registerBackend(backendName, () => {
+      let storedValues: BackendValues = null;
+      return {
+        id: 1,
+        floatPrecision: () => 32,
+        write: (values: BackendValues, shape: number[], dtype: DataType) => {
+          const dataId = {};
+          storedValues = values;
+          return dataId;
+        },
+        read: async (dataId: object) => storedValues,
+        dispose: () => null,
+        disposeData: (dataId: {}) => null,
+      } as TestStorage;
+    });
+    tf.setBackend(backendName);
+
+    const a = tf.tensor1d([1, 2, 3]);
+    expectArraysClose(await a.data(), [1, 2, 3]);
+    a.dispose();
+
+    tf.removeBackend(backendName);
   });
 });
